// types.ts
type UserDbRow = Database["public"]["Tables"]["users"]["Row"];
export interface User extends UserDbRow {}

type ClusterDbRow = Database["public"]["Tables"]["clusters"]["Row"];
export interface Cluster extends ClusterDbRow {}
// ...
type FeedbackDbRow = Database["public"]["Tables"]["feedback"]["Row"];
export interface Feedback extends FeedbackDbRow {}
// ...
export enum ViewName {
  MainMenu = 'Main Menu',
  Dashboard = 'Dashboard',
  TourismCluster = 'Tourism Cluster',
  TourismMapping = 'Tourism Mapping',
  ManageMyClusters = 'Manage My Clusters',
  GrantApplications = 'Grant Applications',
  TourismStatistics = 'Tourism Statistics',
  EventsCalendar = 'Events Calendar',
  AIPlanner = 'AI Planner',
  UserManagement = 'User Management',
  WebsiteManagement = 'Website Management',
  SystemFeedback = 'System Feedback',
  Settings = 'Settings',
  HelpCenter = 'Help Center',
  TestingDashboard = 'Testing Dashboard',
}